backend/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.30
scikit-learn==1.5.0
openai==2.33.0
pandas==2.2.2
joblib==1.4.2
flask-cors==6.0.2
flask-session==0.8.0